Еще рефактора
